Remove HDOP
For now let's not parse the hdop value and put it into the accuracy field. Accuracy is assumed to be an accuracy in metres. HDOP is a unitless indicator of the accuracy. An accuracy of 3 metres is good, while an HDOP of 3 is quite bad.

Estimating an accuracy in metres form HDOP is possible, but that should be left to the (user) payload decoder to do, and keep this parser clean without any magic.
Ref: https://www.gps-forums.com/threads/roughtly-converting-dop-to-metric-error.40105/#post-206314